requirements.txt
pypdf>=4.0.0
ragas>=0.1.7
openai>=1.10.0
python-dotenv>=1.0.0
asyncio>=3.4.3
pandas>=2.1.0
tqdm>=4.66.0
jinja2>=3.1.0
langfuse>=4.0.0
